requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
astroid==3.2.4
    # via pylint
async-timeout==4.0.3
    # via
    #   app (setup.py)
    #   redis
attrs==24.1.0
    # via pytest-docker
autoflake==2.3.1
    # via app (setup.py)
bandit==1.7.9
    # via app (setup.py)
black==25.1.0
    # via app (setup.py)
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   uvicorn
coverage==7.6.1
    # via
    #   app (setup.py)
    #   coverage-badge
coverage-badge==1.1.2
    # via app (setup.py)
cryptography==44.0.2
    # via
    #   jwcrypto
    #   oic
    #   pyopenssl
    #   types-pyopenssl
    #   types-redis
debugpy==1.8.12
    # via app (setup.py)
defusedxml==0.7.1
    # via oic
dependency-injector==4.41.0
    # via app (setup.py)
dill==0.3.8
    # via pylint
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
faker==33.3.1
    # via app (setup.py)
fastapi==0.115.12
    # via app (setup.py)
freezegun==1.5.1
    # via app (setup.py)
future==1.0.0
    # via pyjwkest
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.28.1
    # via app (setup.py)
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
inject==4.3.1
    # via app (setup.py)
isodate==0.6.1
    # via python3-saml
isort==5.13.2
    # via pylint
jinja2==3.1.6
    # via app (setup.py)
jwcrypto==1.5.6
    # via app (setup.py)
lxml==5.3.0
    # via
    #   app (setup.py)
    #   python3-saml
    #   xmlsec
mako==1.3.9
    # via oic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mccabe==0.7.0
    # via pylint
mdurl==0.1.2
    # via markdown-it-py
mirakuru==2.5.2
    # via pytest-redis
mypy==1.11.1
    # via app (setup.py)
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
oic==1.7.0
    # via pyop
packaging==24.2
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
pbr==6.0.0
    # via stevedore
platformdirs==4.2.2
    # via
    #   black
    #   pylint
pluggy==1.5.0
    # via pytest
port-for==0.7.2
    # via pytest-redis
psutil==6.0.0
    # via mirakuru
pycparser==2.22
    # via cffi
pycryptodomex==3.21.0
    # via
    #   oic
    #   pyjwkest
    #   pyop
pydantic==2.8.2
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.20.1
    # via pydantic
pydantic-settings==2.3.4
    # via oic
pyflakes==3.2.0
    # via autoflake
pygments==2.18.0
    # via rich
pyjwkest==1.4.2
    # via oic
pylint==3.2.6
    # via app (setup.py)
pynacl==1.5.0
    # via app (setup.py)
pyop==3.4.1
    # via app (setup.py)
pyopenssl==25.0.0
    # via app (setup.py)
pytest==8.3.2
    # via
    #   app (setup.py)
    #   pytest-asyncio
    #   pytest-docker
    #   pytest-mock
    #   pytest-redis
pytest-asyncio==0.24.0
    # via app (setup.py)
pytest-docker==3.1.1
    # via app (setup.py)
pytest-mock==3.14.0
    # via app (setup.py)
pytest-redis==3.1.3
    # via app (setup.py)
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   freezegun
python-dotenv==1.0.1
    # via pydantic-settings
python-multipart==0.0.19
    # via app (setup.py)
python3-saml==1.16.0
    # via app (setup.py)
pyyaml==6.0.1
    # via bandit
redis==5.2.1
    # via
    #   app (setup.py)
    #   pytest-redis
requests==2.32.3
    # via
    #   oic
    #   pyjwkest
rich==13.7.1
    # via bandit
six==1.16.0
    # via
    #   dependency-injector
    #   isodate
    #   pyjwkest
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
stevedore==5.3.0
    # via bandit
tomli==2.2.1
    # via
    #   autoflake
    #   black
    #   mypy
    #   pylint
    #   pytest
tomlkit==0.13.0
    # via pylint
types-cffi==1.17.0.20250319
    # via types-pyopenssl
types-pyopenssl==24.1.0.20240722
    # via types-redis
types-python-dateutil==2.9.0.20241206
    # via app (setup.py)
types-redis==4.6.0.20241004
    # via app (setup.py)
types-requests==2.32.0.20250306
    # via app (setup.py)
types-setuptools==71.1.0.20240818
    # via types-cffi
typing-extensions==4.12.2
    # via
    #   anyio
    #   astroid
    #   black
    #   faker
    #   fastapi
    #   jwcrypto
    #   mypy
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   uvicorn
urllib3==2.2.2
    # via
    #   requests
    #   types-requests
uvicorn==0.34.0
    # via app (setup.py)
xmlsec==1.3.14
    # via
    #   app (setup.py)
    #   python3-saml

# The following packages are considered to be unsafe in a requirements file:
# setuptools
